refactor e2e-tests
1:1 mapping of canisters to tests
