+++ e2e-tests/src/bin/call_struct.rs
