+++ e2e-tests/src/bin/canister_info.rs
