+++ e2e-tests/src/bin/simple_kv_store.rs
